Fix state so it doesnt exhaust the input by calling popFront too many times.
Because the lexer is a while(input is not empty) loop, if you empty
the input in one state ( putting it all into the char buffer,
and return a new state, the new state wont get
entered, because the lexer think's you're done.

Instead of exhausting the input, don't do that, and let the new pop the
last character off the input.

We should add contracts to every state to make sure none of them
do bad things and leave the input in a reasonable state, and consistent
for the next state.